Fix test package being required for built in fuzz test
Whilst most users whom do testing will probably be using actual tests
rather than simply the fuzzer, we should not be requiring a test package
for a test that doesn't use a test package. This commit changes our
validation and schedule-run logic to not check for and use a test
package on the BUILTIN_FUZZ test.